\name{PBImisc-package}
\alias{PBImisc-package}
\alias{PBImisc}
\docType{package}
\title{
PBI misc
}
\description{
A set of datasets used in my classes or in the book ,,Modele liniowe i mieszane w R, wraz z przykladami w analizie danych''
}
\details{
\tabular{ll}{
Package: \tab PBImisc\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2011-01-01\cr
License: \tab GPL-2\cr
}
}
\section{General Description}{
A set of datasets some of them are my original ones, some are taken from other packages of literature.
}
\author{
Przemyslaw Biecek

Maintainer: You should complain to Przemyslaw Biecek  <przemyslaw.biecek@gmail.com>
}
\references{
The book ,,Modele liniowe i mieszane w R, wraz z przykladami w analizie danych''
}
\examples{
# here you will find some examples
#
}
