\name{PBImisc-package}
\alias{PBImisc-package}
\alias{PBImisc}
\docType{package}
\title{
Set of supplementary datasets and functions
}
\description{
 A set of datasets and functions used in the book
,,Modele liniowe i mieszane w R, wraz z przykladami w analizie danych''
}
\details{
\tabular{ll}{
Package: \tab PBImisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-02-15\cr
License: \tab GPL-2\cr
}
}
\section{General Description}{
A set of datasets some of them are my original ones, some are taken from other packages of literature.
}
\author{
Przemyslaw Biecek

Maintainer: You should complain to Przemyslaw Biecek  <przemyslaw.biecek@gmail.com>
}
\references{
Przemyslaw Biecek ,,Modele liniowe i mieszane w R, wraz z przykladami w analizie danych'' 2013, Wydawnictwo PWN
}
\examples{
# here you will find some examples
#
}
